[EM-37] move models to dedicate directory
