+++ src/main/java/it/andmora/expensesmonitor/domain/PaymentDao.java
@@ -1,6 +1,6 @@
 package it.andmora.expensesmonitor.domain;
 
-import it.andmora.expensesmonitor.domain.entity.Payment;
+import it.andmora.expensesmonitor.domain.model.Payment;
 import reactor.core.publisher.Mono;
 
 public interface PaymentDao {
